Enhanced wrong logic about when we need to check for transaction status on the remote node before staring to copy a set on subscribe.
There is still something wrong here. But on a larger scale. If we
are listening for events from the origin directly, but subscribe via
another slave as provider, the event will most likely be queued by
the origin-listener first. Later on, we ignore duplicate events and
thus, the event will never arrive through the provider.

This means that the remote listener must get more intelligent about
event types and queueing.

Jan