Additions due to the "productionizing" of the 2nd Generation Slon Watchdog.
Also, very importantly, the Slonik script generators no longer
submit slonik output directly to the slonik instance.  You
should REALLY review the slonik before submitting it.  Otherwise
there are plenty of opportunities for this to be a big set of
"foot guns."

And so you are free to, if you wish, submit the output of scripts
directly to slonik by piping them to slonik.