Added in node_is_subscribing() function to make watchdog less aggressive about restarting slon nodes
If a node is in the process of being subscribed, then you can get
long-running COPY requests that run for hours (this happens if it's a
very large database).

slon_watchdog2.pl therefore has two criteria to use to "throttle" its
slon; thus if there has not been a local event on the node in the
interval $killafter, then it will kill (and ultimately restart) the slon
UNLESS a subscription is in progress.