Added a dynamic SYNC group "windowing" scheme, along with docs and configuration option code.
This allows defining a target "desired_sync_time", defaulting to 1
minute, where, if replication is behind, slon will gradually increase
sync group sizes to try to get the expected time for the group to
take that period of time.

It backs off IMMEDIATELY, if there is a large or long running SYNC,
and starts at 1, and increases group sizes by the greater of 1 and
10% of the last group size.  Once the node gets up to date, that
also causes it to back off to sync sizes of 1 or 2.

This makes it reasonable to have the maximum sync grouping set rather
high, as the groups will only get large gradually, in keeping with
how the system is performing.