Use fixed paths for altperl tools configuration and library files.
The configuration file defaults to the "sysconfdir" argument to
configure (/usr/local/etc by default), and has been renamed to
slon_tools.conf, which seems to go along with the naming convention of
the configuration files in /usr/local/pgsql/data.

The slon-tools.pm library is installed in the pglibdir path.

I've also updated the relevant documentation to note the slon.env name
change.