First cut on the log shipping mechanism.
The tools/slony1_dump.sh script can be used against a subscriber
to get a snapshot of a replica's user data plus minimal status
information.

If slon is started with -a <archivedir>, it will output an sql
script for every sync processing, containing statements that replicate
the changes plus keep the status information up to date. Note that these
scripts will have the same "grouping" of SYNC events that the replica
used.

Next steps:
    Add output of archive files by slon for subscribing, unsubscribing
	and merging of sets and for ddl script executions.

Jan