Added support for sets originating on different nodes.  The default remains the $MASTERNODE, but you can now add an "origin" attribute in the configuration file to designate that the set originates on some other node.
Removed the requirement that each node have a parent.  This is now an
optional attribute in the configuration file -- otherwise, it is
assumed that the node can contact any other node that doesn't have a
parent, plus any node where it is the parent.

Modified init_cluster.pl so that it only writes "store path" commands
for paths that exist.

Fixed a couple of bugs in init_cluster.pl that were preventing it from
choosing the most efficient listen path between a given pair of nodes.