Changes to address problems found when running multiple Slony-I clusters against a single backend...
The problem is that if two slons are invoked at almost exactly the same
time, then the cleanup threads will be pretty nearly in phase with one
another. Which gives a (regrettably) excellent chance that both will try
doing an ANALYZE of pg_catalog.pg_listeners concurrently. They'll both
try to commit statistics for it, and one slon will therefore fail.

1.  At startup time, a random "bias" is calculated, and the slon
    sleeps up to an extra 100000ms (e.g. - 100s) between iterations

2.  A further "fuzz" of up to 100s is added

3.  The 7 tables are vacuumed in 7 separate queries

4.  If any of the vacuums fails, this is treated as a SLON_ERROR, not
    as a SLON_FATAL problem requiring that the slon fall over.