From Andreas Pflug <pgadmin@pse-consulting.de>:
In schedule.c and slon.c, pthread_self() result is compared with a
pthread_t variable, which isn't the recommended way to check for
equality; instead pthread_equal should be used. Under win32 this fails,
because pthread_t isn't a simple value. The attached file provides a
patch for this, and should make this portable for all pthread_t platforms.