This patch adds two new functions:
- slon_quote_brute is used in internal functions to add quotes around
identifiers. It replaces all calls to quote_ident in all our functions. This
way we quote every identifier, instead of just the reserved ones.

Pro: we catch every problem with names in Caps and reserved words.
Con: This will increase the size of a row in sl_log_1, as everything is
quoted.

- slon_quote_input is used in internal functions to make sure that user input
is quoted properly.

Known issues:
Indexes with problematic names are not supported yet!


This patch touches a lot of functions, so please test this while we are still
in beta!