Addressing bug report #1323...
There is a function called slon_quote_input() which puts extra "quotes"
into the input string in order to make it a valid fully quoted table
name or namespace name.

The version of the function in Beta5 uses arrays (and is somewhat gross;
I had a difficult time following how it worked) in a way not supported
by PostgreSQL 7.3.x.

In addition to replacing the function, it also turns out to be needful
to use the function in several queries in remote_worker.c where the
names of objects were not being quoted using this consistent approach.

Tested on...
 1. Debian/Linux (7.3.9, 8.0.2, 7.4.8)
 2. AIX (8.0.3)
 3. Solaris (8.0.3)