Resolve bug #1362
Change subscribeset() to verify that the provider is either:
 a) The origin for the set, or
 b) An active forwarding subscriber for the set

The observed result was...

-----------------------------------------------------------------------
I have a cluster replicating several sets, and yesterday I tried to add
yet another node to it and to subscribe the new node to get set5 from
"node3" which was subscribing to the master node "node1".

I noticed that node3 was subscribing with forward=no, so I made a small
slonik script to set subscription to forward=yes. unfortunately I made a
small typo (or maybe a thinko) and also wrote the set id (5) as
provider, so that instead of subscribing to node1, node3 was now
subscribing to node5, which did not even have any of the tables in set5.

Later, when I noticed my mistake, I changed the provider back
to node1, but I'm now missing the changes that had happened in between.
-----------------------------------------------------------------------