Initial Win32 port, from Magnus Hagander, with minor tweaks by Dave Page. Needs service and eventlog code yet, but works nicely from the command line :-)
Changes:
* Added "override" to LDFLAGS in the Makefile. I am no makefile guru, so
I don't really know why :-) But it was there for CFLAGS, and without it,
the pthreads library simply wasn't added to the build command...

* For win32 only, added mandatory link to "wsock32" winsock functions,
required so we can access TCP/IP functions at all.

* #ifdef out the whole "watchdog process" for win32. Win32 does not have
signals, so we don't need a special process to catch and channel those.
And we expect the service control manager to deal with restarts as
required (when it comes).

* As a consequence of win32 not having signals, change exits to just do
exit() instead of sigterm:ing the own process. This includes modifying
slon_abort() and slon_restart() (_restart seems never to be used?)

* Import pgpipe implementation from postgresql backend, and change all
read/write to piperead/pipewrite - same as in the pg backend. This is
rquired because on win32 you can't "select()" on a pipe. The pgpipe
implementation uses TCP sockets instead.

* Remove the sched_shutdown() code per my previous mail. Not strictly a
part of the win32 work...

* Start and stop winsock as appropriate.