Per Ian Burrell - add support for slon-tools.pm $SYNC_CHECK_INTERVAL
It was suggested that we increase the sync check interval (-s).  We
are using slon_start to the daemons.  The value (-s 1000) is hardcoded
in slon-tools.pm.  Instead of editing that file, I added support for
setting the value in the conf file.

The appended patch adds a new config variable, $SYNC_CHECK_INTERVAL,
to slon_tools.conf and uses it in start_slon.  It maintains the
current defaultt of 1000, although shouldn't it have the true default
of 1000.  It also does some refactoring of the construction of the
command line.