Add "lag_interval" option to slon
This allows a slon to have, passed to it, a time interval.  The slon
will then ignore events until they have become older than that interval.

This allows the Gentle User to set up a subscriber that will steadily
lag by 1 hour, so you could query an old state of the database.