Log Message: ----------- Fix UTF-8 string truncation bug
This was reported by Ian Burrell <ianburrell at gmail.com>
along with the suggested patch...

-----------------------------------------------------------------------
"I didn't see any response to the bug I found with slon truncating
strings with UTF-8 characters in them.  The truncated strings were
being written to the sl_log_1 table.  This was causing the SYNC to
fail on the slave because of constraints violations.  It was also
corrupting data.  The slon_quote_literal shortens the copied string by
one byte for every multi-byte character."
-----------------------------------------------------------------------

I requested some details on how to verify testing of this; the major
bit is to create the database with --encoding UNICODE, which then
allows a string to contain Unicode values.

I used this to create the new test script test_F_utf8...

Without the suggested patch, the test script generates the following
difference file (on my system, based on my password file :-)); note
that on the subscriber, the "e" at the end of each row has been
truncated.

1,2c1,2
<  id |          string
< ----+---------------------------
---
>  id |          string
> ----+--------------------------
7,31c7,31
<   5 | sys - Ã¡ - Unicode
<   6 | sync - Ã¡ - Unicode
<   7 | games - Ã¡ - Unicode
<   8 | man - Ã¡ - Unicode
<   9 | lp - Ã¡ - Unicode
<  10 | mail - Ã¡ - Unicode
<  11 | news - Ã¡ - Unicode
<  12 | uucp - Ã¡ - Unicode
<  13 | proxy - Ã¡ - Unicode
<  14 | www-data - Ã¡ - Unicode
<  15 | backup - Ã¡ - Unicode
<  16 | list - Ã¡ - Unicode
<  17 | irc - Ã¡ - Unicode
<  18 | gnats - Ã¡ - Unicode
<  19 | nobody - Ã¡ - Unicode
<  20 | Debian-exim - Ã¡ - Unicode
<  21 | gdm - Ã¡ - Unicode
<  22 | saned - Ã¡ - Unicode
<  23 | sshd - Ã¡ - Unicode
<  24 | postgres - Ã¡ - Unicode
<  25 | postfix - Ã¡ - Unicode
<  26 | messagebus - Ã¡ - Unicode
<  27 | chris - Ã¡ - Unicode
<  28 | ntp - Ã¡ - Unicode
<  29 | hal - Ã¡ - Unicode
---
>   5 | sys - Ã¡ - Unicod
>   6 | sync - Ã¡ - Unicod
>   7 | games - Ã¡ - Unicod
>   8 | man - Ã¡ - Unicod
>   9 | lp - Ã¡ - Unicod
>  10 | mail - Ã¡ - Unicod
>  11 | news - Ã¡ - Unicod
>  12 | uucp - Ã¡ - Unicod
>  13 | proxy - Ã¡ - Unicod
>  14 | www-data - Ã¡ - Unicod
>  15 | backup - Ã¡ - Unicod
>  16 | list - Ã¡ - Unicod
>  17 | irc - Ã¡ - Unicod
>  18 | gnats - Ã¡ - Unicod
>  19 | nobody - Ã¡ - Unicod
>  20 | Debian-exim - Ã¡ - Unicod
>  21 | gdm - Ã¡ - Unicod
>  22 | saned - Ã¡ - Unicod
>  23 | sshd - Ã¡ - Unicod
>  24 | postgres - Ã¡ - Unicod
>  25 | postfix - Ã¡ - Unicod
>  26 | messagebus - Ã¡ - Unicod
>  27 | chris - Ã¡ - Unicod
>  28 | ntp - Ã¡ - Unicod
>  29 | hal - Ã¡ - Unicod

With the patch, the test runs and finds no differences...