Per Ian Burrell...
"I just ran into a bug which crashes slonik when it is passed a file
which doesn't exist.  It looks like it was fixed on HEAD as part of
revision 1.43 on src/slonik/slonik.c. There are other changes in that
revision so I have extracted the relevant patch and appended it.  It
looks like it could be applied separately."