Fixed weird RPM build problem. Previously, the spec file under redhat/ caused some problems like error: line ... second %prep.
To fix this, instead of using -ta, I moved spec file and tar.gz into RPM build locations and and performed the RPM building there. This is the safest way. Same problem arises in Apache2.1 RPM series also, so maybe we can consider this as a bug. This little patch took my 5 nights...

rpm: dist
        mv @PACKAGE_NAME@-@PACKAGE_VERSION@.tar.gz /usr/src/redhat/SOURCES
        mv @PACKAGE_NAME@.spec /usr/src/redhat/SPECS
        rpmbuild -ba /usr/src/redhat/SPECS/@PACKAGE_NAME@.spec