Fixed failover processing.
In the case of failover we need the same ACCEPT_SET guarding, so that
the remaining subscribers do not process any events from the backup
node before they have finished all remaining events from the failed
node, that might be queued up on the backup.

ACCEPT_SET must also restart slon in order to guarantee that the new
configuration is loaded.


Jan