On first SYNC, need to compress log_actionseq not in (possibly huge list) into a set of "not (log_actionseq between a and b or log_actionseq between c and d ...)"
If the SUBSCRIBE_SET event runs for a very long time, there can be a whole lot
of events on the go between the time it starts and that first SYNC.  If so,
the list of log_actionseq values can consist of thousands if not greater
orders of magnitude of items.

compress_actionseq() takes the "huge list" (a string of comma delimited values)
and has a state machine to parse out the numbers, in order.

Each number is compared to the current 'range'...
- If it's inside or extends a range, we continue working on a "between a and b" clause
- The value might be a singleton, thus adding a "log_actionseq <> x" clause

What generally happens if the event runs for a long time is that there will
be some really enormous sequences of consecutive values.  Happily, the
new function shrinks that into "between a and b" (where b = a + 50000), which
means we have a query that's 4K in size rather than 40MB in size.

Hannu Krosing added logic at the end that handles the case where there
was NOTHING in the set of log_actionseq values.