Long cluster name causes cluster to not operate properly (ID: 1436)
"In set creation I allowed slony1 to create unique indexes on a handful
of unkeyed tables. When the cluster name was long, the complete name
it uses for those columns was truncated. The create set script/process
did not detect this condition and thus all subsequent operations on
those tables failed. I also had to manually drop those columns, as the
uninstall process failed as well."

Changed node creation to check for this, comparing cluster name size
to NAMEDATALEN.