Support case-folding - Philip Yarra <philip at utiba.com>
I figure the easiest way to explain what I propose wrt. case-folding is to
supply a patch. No changes to the DB functions, just a small change in
slon_tools.pm and slonik_create_set.pl

If you define:
foldCase => 1,
in slon_tools.conf, object names (including schema - wasn't sure if that was
wise or not) will be folded to lower case.

The default is 0, so no change to existing behaviour unless you specifically
say so.

I also changed 'table1' -> 'TABLE1' in the sample conf to illustrate the
difference.