FEATURE REQUEST #1443 - add cluster name to "Node is uninitialized" error message
During a switchover test, we got the following error message:

ERROR: Slony-I: Node is uninitialized

It happened because a table we were using for a heartbeat monitor
erroneously had an extra slony trigger on place for a different (no
longer active) slony cluster.

Totally strange circumstance, and totally my fault, but debug time would
have been much reduced if the error message had told us which node in
which cluster was not initialised.

Is there any chance that this information could be added to the error
message in a future release?

marc <NOSPAM> bloodnok.com
---------------------------------------------------
Have added a patch to CVS HEAD that lists the cluster name.
Unfortunately, the node number is not available at the point in the code
where this error message is issued.

cbbrowne@ca.afilias.info