Revise man.sgml / slony.sgml to have a set of ENTITY values defined so as to allow man.sgml to omit all references to the schema document.
In man.sgml, references to schema objects are turned into simple
literals (or <function> entities).  In slony.sgml, those references
have the 'more sophisticated' references to objects in schemadoc.xml

The point here is to have man.sgml be a feasible totally-valid SGML
document that we can use to generate manual pages.