Implementation of "SWITCH LOG (id = <nodeid>)".
The slonik command SWITCH LOG will call the stored procedure

	logswitch_start()

which will adjust the sl_log_status sequence or bail out with an
ERROR if the previous logswitch did not complete yet. The slon
cleanup thread will call the stored procedure

    logswitch_finish()

which will check if the old log is logically empty (if log_status
is 2 or 3) and if so, truncate the old sl_log_? and adjust sl_log_status
to the final value completing the switch.

This patch also includes the completion of functionality in the log
trigger and the remote worker thread to actually use sl_log_2.

Jan