Bug #1519 - Should explicitly check slony version on store node command
Add a test function early in slony1_funcs.sql that checks the
version information in the SPI stored procedure against a macro
expansion of @MODULEVERSION@ done by slonik.

The function aborts with an explanatory error message if those
versions do not match.

This will cause a much earlier, explained, abort of STORE
NODE/INITIALIZE CLUSTER if there is a version mismatch between slonik
and the stored functions/schema.