Version checking code in slonik.c broke in the way it looked for elderly versions; apparently a < and > got swapped...
Fixed that, as well as adding to the "your version of PostgreSQL is too old"
message some indication as to how bad that situation is.

For instance, if you're on PG 7.3, then Slony-I 1.1.5 is still an option.

But if you're on < 7.3, Slony-I never was an option (at least not for
those that lack near-PG-core-level hacking-on-slon abilities...)