Partial sl_log_? indices support
Add a function, addPartialLogIndices(), which adds missing partial indices
against the unused sl_log_? table, and drops any that are no longer
needed.  (Needed ==> "Node # is an origin for a set")

This function is run in various places that touch set origins so that
the indexes will, over time, be available on both tables, as log
switches take place.