Three changes:
1.  Allocate an extra entry in sync_group - SlonWorkMsg_event *sync_group[MAXGROUPSIZE+1];

That way, we can get the value of sync_group_size all the way to
MAXGROUPSIZE and not blow out the end of the array.

2.  Add into the while loop that increments sync_group_size a check that we haven't
exceeded MAXGROUPSIZE.

That way we won't blow past the end of the array if next_sync_group_size gets set too
high.

3.  Logic for processing return code from fclose() treats "nonzero" as the error
condition as per H&S Section 15.2