This is the first step towards Slony-I version 2.0.
It drops all support for databases prior to Postgres version 8.3.
This is required because we now make use of new functionality in
Postgres, namely the trigger and rule support for session replication
role. As of now, every node (origin/subscriber/mixed) can be dumped
with pg_dump and result in a consistent snapshot of the database.

TODO:
  - Run DDL scripts as "local" replication role. Currently DDL
    scripts don't work at all.
  - Get rid of STORE TRIGGER and force users to make use of 8.3's
	advanced trigger and rule configuration.

Jan