From: Bernd Helmle <mailings@oopsware.de>
--On Donnerstag, Oktober 25, 2007 13:11:49 -0700 slony@estrider.com wrote:
>    There is a flaw in the slonik_build_env.pl script. The print test
> should be >=1 and not > 1. It currently will not print databases where
> there are only single tables or sequences.
>

You're right. I've created a patch with your changes plus a fix to
repair the -schema option, which was broken, too, and cc'ed
slony1-patches so someone can review and apply it.

Note that slonik_build_environment is called slonik_build_env only,
the former was removed.

--
 Thanks

                    Bernd
-------------------------------------------------
Looks good to me - applied to HEAD branch...