Per bug #38 - reported by Peter Eisentraut
The way the function start_slon in slon-tools.pm assembles the command line
creates a command that looks like this:

slon ... 2>&1 >logfile &

This ends up dumping the error messages on the console (or wherever) and only
stdout to the log file.  What you want is this:

slon ... >logfile 2>&1 &

The patch fixes this.