Remove code that generates + listens to LISTEN/NOTIFY events for   Events + Confirms
remote_listen.c now functions solely via polling, which we already know
works fine, as we have had a polling mode for quite a while in the 2.0
code tree.

The slon _Restart LISTEN/NOTIFY has been left alone; making sure that this
is handled rightly will be left as a further exercise to be handled
independently.