Changed ducttape test 2 to use "wait for event" after subscriptions in order to subscribe node 3 as soon as possible, which is when node 2 has finished the copy set operation and has done it's very first sync processing.
This revealed a problem in the remote worker. Messages that forward
confirmations need to be inserted at the head of the internal message
queue in order to get processed in case the node is waiting for the
data provider to catch up before starting to copy set in a cascaded
setup.

Jan