Fix to bug #69
http://www.slony.info/bugzilla/show_bug.cgi?id=69

The problem is that one of the new functions expects the type, "vactables", to
exist.

The "upgrade functions" function has code that *would* create it, but that
function is not run until *after* the whole slony1_functions.sql file has been
loaded, and that is way too late.

The attached patch creates, runs, and drops, a temporary function that adds the
type should it be missing.

I have run this through regression test #1 on the HEAD branch, which includes a
run of slonik "UPDATE FUNCTIONS."