Applied patch from bug #67 http://www.slony.info/bugzilla/show_bug.cgi?id=67
-----
I attached a small patch against slony1_funcs.c which checks for the 0
byte at the end of the attkind string and stops the iteration. This
patch also adds some error checking in the loop beginning in line 648.
Of course the problem can be removed by not trimming the string in the
determineAttkindUnique(text, name) function, but I think that the end
of a zero-terminated string should be checked in any case.
-----

Thanks to Adam Buraczewski...