Aleksander Kmetec submitted this patch which addresses a race condition potentially effecting long time locking or data loss on TRUNCATE.
Dataloss bug, described here:
http://lists.slony.info/pipermail/slony1-general/2009-July/009663.html

Statement blocking bug, described here:
http://lists.slony.info/pipermail/slony1-general/2009-July/009664.html

Patch is to preface TRUNCATE logic with an attempt (in an exception
block) to acquire an exclusive lock on the apropos sl_log_* table.
The NOWAIT option raises a (caught!) exception if there are any other
processes holding onto the table.