Remove direct uses of yyleng and instead use the function yyget_leng() More recent versions of flex have changed yyleng from an int to a size_t. By using the function we will let the compiler do the conversion if required on the function result.
We also add an explicit object for scan.o instead of including it at the
bottom of parser.y because we need to generate and include a header for
the scanner (scan.h) so we have the explicit(and correct) definitions of
the yyyget_leng() function.

A autoconf check to ensure that a version of flex with yyget_leng() is
available has also been added.