commit e86d296ae8c739018100306bcebeded80525f265 seems to have been a merge or replace of the master branch into REL_2_0_STABLE.
This is not what should have been intended as it erased numerous changes
that were only in the 2.0 branch that should be merged into master.

This commit will put those changes back into 2.0, it undoes parts of
e86d29 but leaves the CVS tags removed.

This commit also undoes some changes made in master (mostly documetnation)
that maybe should go in 2.0. If that is the case they will need to be
put back as seperate commits

This commit was prepared by going back to before e86d29 and redoing the
tag removal and .gitignore changes then merging in the other changes done to
REL_2_0_STABLE after.  This patche was then prepared to make the current
REL_2_0_STABLE look like the result.