Remove direct uses of yyleng and instead use the function yyget_leng() More recent versions of flex have changed yyleng from an int to a size_t. By using the function we will let the compiler do the conversion if required on the function result.
We also add an explicit object for scan.o instead of including it at the
bottom of parser.y because we need to generate and include a header for
the scanner (scan.h) so we have the explicit(and correct) definitions of
the yyyget_leng() function.

A autoconf check to ensure that a version of flex with yyget_leng() is
available has also been added.
(cherry picked from commit b44f967e2c62c60f5a2448ce1c4d5afd878922bf)

Conflicts:

	RELEASE
	configure
	configure.ac
	src/slonik/Makefile
	src/slonik/parser.y
	src/slonik/scan.l
	src/slonik/slonik.h
	src/slony_logshipper/Makefile
	src/slony_logshipper/parser.y
	src/slony_logshipper/scan.l