Bug 42:
The problem is that postgresql_autodoc produces XML but the documentation is
processed using SGML.  SGML does not process the empty tag notation <xref .../>
correctly.  Instead it should be <xref ...> without slash.  I have attached a
patch that fixes up the XML to be valid SGML so the documentation comes out
right.

Patch by Peter Eisentraut