Bug #115
Cause the parser to generate an ERROR if a WAIT FOR EVENT
occurs inside of a TRY block.

In reality, the wait itself isn't the problem. It is the combination
of creating the event (sync, store whatever) and waiting for it
inside the same try block. Try blocks don't commit until they end,
so the event never gets committed and logically, the wait must wait
forever.