Bug 149
Fix for when local listen fails to start properly. It was observed that slon can wait for the local listen cond to be set where the local listener thread stops on an error. In this case we want to signal the mutex but flag the local listener as not having started. The main slon thread should then abort.
(cherry picked from commit ba234cfe686f33fd10d2af1c557af66f8ec17b02)