Changes to address some issues found in testing FAILOVER in an auto-wait for world.
When the FAILOVER_SET command is submitted only slonik knows
which node it was submitted on (the most ahead node).
The system needs to wait for that FAILOVER_SET to be confirmed
on all nodes before proceeding.  Only slonik can do this
wait for because the event has ev_origin=failed node but
shows up in the sl_event of the most ahead node. Also the
event number is faked by slonik and is 'lost' after
the command finishes.