Disabling some wait for statements in the test since in an auto-wait for world they should not be required.
Adding explicit slonikSync() commands before the compareDb operations since
there is nothing else that ensures the slaves are caught up (since we removed
the explicit wait commands and an auto-wait would only happen before the
next slonik command that requires it).

This test still fails. What seems to happen is that the SUBSCRIBE SET 3->5
gets an event sequence id such as (1,80) but when FAILOVER executes
it only sees event (1,79) on any of the other nodes so it fakes the
event (1,80) to be the FAILOVER_SET.  However the slon on node 5
pulls the (1,80) SUBSCRIBE SET before node 5 sees the FAILOVER_SET.
Node 5 thinks it has seen the event (1,80) when it has seen the
'wrong' version of this event.

Fixing this is beyond the scope of auto-wait for.