Adding test suites based on the clustertest-framework.
clustertest-framework is a testing framework developed for testing
slony that can be found at
https://github.com/clustertest/clustertest-framework/

This commit adds two test suites:

clustertest/regression - is a port of the shell script based regression
tests from the slony1-engine/tests directory.

clustertest/disorder - is a new set of tests for testing slony in the context
of a simulated order entry system.

The tests are based on version 1.0.0  of the clustertest-framework.

The disorder simulated business transactions were written by Jan Wieck.
The scripts that control the transaction flow and slony commands where
written by Steve Singer as was the javascript port of the existing
regression tests.
Chris Browne added documentation, some additional tests, review, testing
and logging improvments.