Provide a way of resetting the per session datastructure that stores the cluster status in the C functions.
cloneFinish() changes the sequence sl_local_node_id but the value
is cached in the C functions per session.  If getLocalNodeId()
has already been called in the current session cloneNode will
do the wrong thing because subsequent calls to getLocalNodeId()
will return the old value.

With the auto-wait-for changes slonik typically calls
getLocalNodeId() before submitting commands this exposes
this bug.

Here we provide a C function that resets the cluster status
and have the cloneNode function call this after changing
the sl_local_node_id value.