Make sl_event_lock locking work with try blocks.
The LOCK sl_event_lock must be before any queries/SQL in the
transaction.  In a try lock where multiple slonik commands
are done as part of the same transaction this causes issues.
The solution to this is that if in a try block then obtain
sl_event_lock when a transaction on the dbconn is opened