Automatic Wait For.
This is a merge of the auto_wait_for feature branch.

This change modifies slonik so that it will (in normal circumstances)
automatically wait for events to be confirmed before proceeding in
cases where it detects that this is required.

See the documentation changes included in this patch for more details.