Bug 204.
The commit 7970189ec93faf3ee71cf49e97529980bbd219aa item (1) says
it removed a duplicate query. That query in fact is/was slightly
different than the one above and should not have been removed.

FAILOVER to a non direct subscriber leaves sl_subscribe with
entires listing the backup node as both a provider and a receiver
of the replication set. This commit should revert that regression.