Failover fixes: 1) The backup node might not be a direct subscriber of the failed node.  If so we don't want to use the 'only direct subscriber' code path but use a code path that deals with non direct subscribers.
2) Undoing e3313e04357e7c084c4e412c87fe47aa64dc0ea9
If we delete the row from sl_subscribe and the backup node is behind
it will process SYNC events from the most ahead node but it will think
that it has 'no sets need syncing' because the sl_subscribe row that
gets added back in by subscribeSet_int(..) has active=false which
means adjust_provider_info() in the remote_worker will skip setting
up that node as a provider.

3) During a failover it is important that all nodes see the ENABLE_SUBSCRIPTION
for the most ahead node before they see the FAILOVER_SET event.  If a node
receives events in the order of FAILOVER_SET then ENABLE_SUBSCRIPTION the
sl_subscribe entries can be wrong because that node will then have a
sl_subscribe row showing data from the most-ahead-node to the backup-node.
This leaves that node inconsistent with other nodes.

Accomplishing (3) meant that subscribeSet_int() needs to return an event id
if it creates an ENABLE_SUBSCRIPTION event and slonik needs to wait for
all remaining nodes to process that event.   This change probably can't
be backpatched to 2.0 or 2.1 because without the multi-node failover syntax
slonik would be waiting for the other failed nodes.