Change format of the sl_log_1/2 tables so that they contain the table name and schema as separate fields. Change the column names and values from a rudimentary SQL statement into a text array.
Change the transport to use COPY protocol and move the part that does
the updates to user tables into a trigger that fires on sl_log_1/2.

Move DDL from sl_event into a separate sl_log_script table that is
pulled into the log selection so that schema changes flow at the
correct time in between data updates, not as separate events.