Significant changes to failover.
We now have explicit support for having multiple nodes FAILOVER at the same time.
The FAILOVER command will take a list of these nodes.

The DROP NODE command can also take a list of multiple nodes that get dropped
at once.

The FAILOVER code changes with this commit.
You can now only failover to nodes that are a failover candidate
as listed in the sl_failover_targets view. This nodes must have paths
to all the nodes the failed node has paths to and be a direct
subscriber to all sets the failed node is an origin for.  See
the patch + documentation changes for a complete list of restrictions.
These restrictions were put in place to avoid complications in
getting a complete path+listen network where a non-direct subscriber
might be the most ahead node.

The FAILOVER command now promotes the most ahead node to be a master
and uses MOVE SET to make the desired backup node the new master.

The FAILOVER_NODE event replaces the FAILOVER_SET event and handles all
sets on that node.  The event now comes from the most-ahead node using
the most ahead node's event_id events from the old origin are no longer faked.
This works because of the restrictions discussed above.