Remerging bug 248.
This was partially unmerged as part of the COPY protocol commits

Stop tracking the sequence name on log shipping targets.
Instead include the sequence name in the log shipping files on
sequence updates.  This avoids the issue of us not adding
new tracking table entries on an enable subscription.

Add an OMIT COPY option to the script to allow
a log shipping target to be upgraded to a new slony version
without reloading all the data.  During an outage you
would drop the $clustername schema from the target
then run slony1_dump.sh with -omit_copy to reinit
the node.