The tutorial in the documentation has this snippet: 8<-----------------     #--     # init the first node.  Its id MUST be 1.  This creates the schema     # _$CLUSTERNAME containing all replication system specific database     # objects. 8<-----------------
That hasn't been true, for a while, and, in reviewing the material in the
howto directory, that's all pretty obsolete, and removal-worthy.  Anything
that *is* valid has been drawn into the admin guide, and this material has
been getting steadily less valid over time due to functionality changes.

Also removed the "MUST be 1" from the tutorial in the admin guide, which
was referenced from the older basic tutorial.  And removed howto material
from RPM spec file