Fix for bug282.
There was a problem with the new COPY protocol when
a data type or domain used in a replicated column
is not in slon's search_path. SPI does not provide
a mechanism to get the namespace name of a columns
data type.

Instead of adding explicit type casting to the apply
queries and handing the data in as TEXT Datums, we
now use the same technique that PL/pgSQL uses at
at least since 8.3 and convert the TEXT datums into
the requested data type ourselves.